languages test (without sensors)
