+++ suites/languagesTest.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="Languages suite">
+    <test name="Spanish language test">
+    <parameter name="browserLanguage" value="es"/>
+    <classes>
+        <class name="selenium.DefaultLanguageTest"></class>
+    </classes>
+</test>
+    <test name="Other language tests">
+        <classes>
+            <class name="selenium.LanguagesTest"></class>
+        </classes>
+    </test>
+</suite>
